Update push-to-s3 - move to public requester-pays bucket. (#11)
This update changes the upload bucket for our s3 data with the rationale of (1) ending object expiration and (2) moving to _requester pays_.

__Object Expiration:__
The old bucket had object expiration which caused problems with GitHub's behavior of disabling actions on repositories with limited activity.

__Requester Pays:__
The new bucket is public to logged-in users (no anonymous access) and is configured with the _requester pays_ behavior so that egress charges are billed to the account originating the file egress than to the account that owns the bucket (our account). In practice we expect JCSDA to be responsible for the majority of egress traffic but if this changes we will not be responsible for high utilization billing.